refactor LoginFields component to AuthForm component for greater use
